Feat/vectordb (#4)
* feat(core:vectordb): sync

* feat(core:vectordb): fix types & bugs

* feat(examples:vectordb): add milvus & modify pinecone

+++ packages/sdk/src/VectorDB/VectorDBInstance.class.ts
@@ -120,12 +120,16 @@ export class VectorDBInstance extends SDKObject {
         return results.map((result) => ({
             embedding: options?.includeEmbeddings ? result.values : undefined,
             text: result.metadata?.text,
-            metadata: typeof result.metadata?.user === 'string' ? JSON.parse(result.metadata.user) : result.metadata?.user,
+            metadata: typeof result.metadata === 'string' ? JSON.parse(result.metadata) : result.metadata,
         }));
     }
 
     public async purge() {
         await this.ready;
+
+        if (!(await this.namespaceExists())) {
+            return;
+        }
         await this._VectorDBRequest.deleteNamespace(this._namespace);
     }
 }
